pyproject.toml
[project]
name = "rag-thesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "dotenv>=0.9.9",
    "google-genai>=1.28.0",
]
